Remove arguments conversion to const char * const *
This was an inherently thread-unsafe interface, using mutable data.
Instead allow a more explicit conversion which returns a unique_ptr.
This still points into data not owned by the pointer, but is safe so
long as the object is not modified.

Leaving poor interface in the environment object as this is not an
urgent fix there.